Uploaded Web Analyzer
During the completion of this lab, we noticed that the frequency of words for exercises 4 and 5 do not match with some being of higher frequency than the other. We have determined that this makes sense! Consider the word "olympics". Exercise 4 requires us to find a phrase which means the word "olympic" will still appear in olympics. However in exercise 5, this requires us to find the exact word and involves splitting the test. This is why olympic will not appear in olympics. There might be some variations depending on the text, but we have ensured to implement the methods as requested and ensured 100% functionality. Thank you so much!